Feat: Housing (#302)
* Feat: WIP Home

* clean up

* only pick arrow cubes

* .

* fix me dumb

* entering private houses, nominate house

* HomeManager and reworks

* more stuff

* more stuff

* WoodlandPath template

* merge fix

* templates + invite to home + kick

* fix template

* decoration score, decoration reward, interior desing reward

* more housing stufff

* clean up

* Update ResponseLoadUGCMapPacket.cs

* Revert "Update ResponseLoadUGCMapPacket.cs"

This reverts commit 680b29cc4bfcd89ba8f522fd92b2c51991500aa1.

* using latest PR RandomProvider

* decoration reward timestamp

* reusing code;
fixed replace packet;
started home layout;
added more checks for home owner;

* reverting changes to PlayerStorage

* home instance counter + bug fixes

* clear interior + starting layouts

* layouts

* dotnet-format

* cleanup

* Update ItemParser.cs

* suggestions

* suggestion

* Update DatabaseContext.cs

* Budget handling + packet

Fixes:
- Cubes in Coord 0 weren't loading in Load layout
- Fixed forfeiting plot
- Not longer stuck in build mode if you change maps while in build mode

* suggestion

* fix: faster loading homes

+++ MapleServer2/PacketHandlers/Game/DungeonHandler.cs
@@ -63,7 +63,7 @@ namespace MapleServer2.PacketHandlers.Game
 
         public static void HandleEnterDungeonPortal(GameSession session)
         {
-            int instanceId = session.Player.InstanceId;
+            long instanceId = session.Player.InstanceId;
             DungeonSession dungeonSession = GameServer.DungeonManager.GetDungeonSessionByInstanceId(instanceId);
             if (dungeonSession == null)
             {
